pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["minyoungci <dbssus123@gmail.com>"]
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.155.2"}


[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
